*** empty log message ***

+++ volume_io/Geometry/colour_def.c
@@ -15,23 +15,23 @@ public  Colour  make_rgba_Colour(
 public  int  get_Colour_r(
     Colour   colour )
 {
-    return( colour & 255ul );
+    return( (int) (colour & 255ul) );
 }
 
 public  int  get_Colour_g(
     Colour   colour )
 {
-    return( (colour >> 8ul) & 255ul );
+    return( (int) ((colour >> 8ul) & 255ul) );
 }
 
 public  int  get_Colour_b(
     Colour   colour )
 {
-    return( (colour >> 16ul) & 255ul );
+    return( (int) ((colour >> 16ul) & 255ul) );
 }
 
 public  int  get_Colour_a(
     Colour   colour )
 {
-    return( colour >> 24ul );
+    return( (int) (colour >> 24ul) );
 }
